refactor: Add MoblieLayout.jsx #45
