+++ src/Components/MobileLayout.jsx
@@ -1,0 +1,39 @@
+import React from "react";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
+import CssBaseline from "@mui/material/CssBaseline";
+import Container from "@mui/material/Container";
+import Box from "@mui/material/Box";
+import Copyright from "./Copyright";
+
+const theme = createTheme({
+  palette: {
+    neutral: {
+      main: "#64748B",
+      contrastText: "#fff",
+    },
+  },
+});
+
+function MobileLayout(props) {
+  return (
+    <ThemeProvider theme={theme}>
+      <Container component="main" maxWidth="xs">
+        <Box
+          height="100vh"
+          sx={{
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "center",
+            justifyContent: "center",
+          }}
+        >
+          <CssBaseline />
+          {props.children}
+          <Copyright sx={{ mt: 8, mb: 4 }} />
+        </Box>
+      </Container>
+    </ThemeProvider>
+  );
+}
+
+export default MobileLayout;
